Updated Controllers

+++ src/main/java/com/foundgrounds/photos/PhotoRepository.java
@@ -2,6 +2,8 @@ package com.foundgrounds.photos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhotoRepository extends JpaRepository<Photo, Integer> {
+import com.foundgrounds.photos.Photo;
+
+public interface PhotoRepository extends JpaRepository<Photo, Long> {
 
 }
